Hello/requirement.txt
flask
Flask-RESTful
Flask-SQLAlchemy
PyJWT
flask-jwt-extended